requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.5.0
fastapi==0.85.2
frozenlist==1.3.1
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.2
numpy==1.23.4
orjson==3.8.1
pandas==1.5.1
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.4
yarl==1.8.1
